added facultyRegData, FacultyService, form for marks, faculties show, deleted registration jsp

+++ admissions/src/main/java/ua/kyiv/admissions/dao/FacultyRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ua.kyiv.admissions.domain.Faculty;
 
-public interface FacultyRepository extends JpaRepository<Faculty, Long> {
+public interface FacultyRepository extends JpaRepository<Faculty, Integer> {
 
 }
